add torch impl for compute/clipping kernel

+++ npbench/benchmarks/compute/compute.py
@@ -6,8 +6,8 @@ import numpy as np
 def initialize(M, N):
     from numpy.random import default_rng
     rng = default_rng(42)
-    array_1 = rng.uniform(0, 1000, size=(M, N)).astype(np.int64)
-    array_2 = rng.uniform(0, 1000, size=(M, N)).astype(np.int64)
+    array_1 = rng.uniform(0, 1000, size=(M, N)).astype(np.float64)
+    array_2 = rng.uniform(0, 1000, size=(M, N)).astype(np.float64)
     a = np.int64(4)
     b = np.int64(3)
     c = np.int64(9)
